+++ reactapp/src/App.js
@@ -1,6 +1,5 @@
 import Navbar from './components/Navbar/Navbar';
-import Mainpage from './components/main_test';
-import inputFile from './components/inputFunction';
+import Mainpage from './components/main';
 import './App.css';
 
 function App() {
