+++ reactapp/src/pages/main.js
@@ -2,6 +2,7 @@
 function Main() {
   return (
     <div className="App">
+      asdfasdf
     </div>
   );
 }
